Move `NewBloodPressureSummaryView` to `bloodpressures` package
