+++ app/src/main/java/org/simple/clinic/summary/bloodpressures/NewBloodPressureSummaryViewUpdate.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.summary.bloodpressures.newbpsummary
+package org.simple.clinic.summary.bloodpressures
 
 import com.spotify.mobius.Next
 import com.spotify.mobius.Update
