+++ app/src/main/java/org/simple/clinic/summary/bloodpressures/view/NewBloodPressureSummaryView.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.summary.bloodpressures.newbpsummary.view
+package org.simple.clinic.summary.bloodpressures.view
 
 import android.annotation.SuppressLint
 import android.content.Context
@@ -28,19 +28,19 @@ import org.simple.clinic.router.screen.ScreenRouter
 import org.simple.clinic.summary.PatientSummaryConfig
 import org.simple.clinic.summary.PatientSummaryScreenKey
 import org.simple.clinic.summary.SUMMARY_REQCODE_BP_ENTRY
-import org.simple.clinic.summary.bloodpressures.newbpsummary.BloodPressureClicked
-import org.simple.clinic.summary.bloodpressures.newbpsummary.AddNewBloodPressureClicked
-import org.simple.clinic.summary.bloodpressures.newbpsummary.NewBloodPressureSummaryViewConfig
-import org.simple.clinic.summary.bloodpressures.newbpsummary.NewBloodPressureSummaryViewEffect
-import org.simple.clinic.summary.bloodpressures.newbpsummary.NewBloodPressureSummaryViewEffectHandler
-import org.simple.clinic.summary.bloodpressures.newbpsummary.NewBloodPressureSummaryViewEvent
-import org.simple.clinic.summary.bloodpressures.newbpsummary.NewBloodPressureSummaryViewInit
-import org.simple.clinic.summary.bloodpressures.newbpsummary.NewBloodPressureSummaryViewModel
-import org.simple.clinic.summary.bloodpressures.newbpsummary.NewBloodPressureSummaryViewUi
-import org.simple.clinic.summary.bloodpressures.newbpsummary.NewBloodPressureSummaryViewUiActions
-import org.simple.clinic.summary.bloodpressures.newbpsummary.NewBloodPressureSummaryViewUiRenderer
-import org.simple.clinic.summary.bloodpressures.newbpsummary.NewBloodPressureSummaryViewUpdate
-import org.simple.clinic.summary.bloodpressures.newbpsummary.SeeAllClicked
+import org.simple.clinic.summary.bloodpressures.AddNewBloodPressureClicked
+import org.simple.clinic.summary.bloodpressures.BloodPressureClicked
+import org.simple.clinic.summary.bloodpressures.NewBloodPressureSummaryViewConfig
+import org.simple.clinic.summary.bloodpressures.NewBloodPressureSummaryViewEffect
+import org.simple.clinic.summary.bloodpressures.NewBloodPressureSummaryViewEffectHandler
+import org.simple.clinic.summary.bloodpressures.NewBloodPressureSummaryViewEvent
+import org.simple.clinic.summary.bloodpressures.NewBloodPressureSummaryViewInit
+import org.simple.clinic.summary.bloodpressures.NewBloodPressureSummaryViewModel
+import org.simple.clinic.summary.bloodpressures.NewBloodPressureSummaryViewUi
+import org.simple.clinic.summary.bloodpressures.NewBloodPressureSummaryViewUiActions
+import org.simple.clinic.summary.bloodpressures.NewBloodPressureSummaryViewUiRenderer
+import org.simple.clinic.summary.bloodpressures.NewBloodPressureSummaryViewUpdate
+import org.simple.clinic.summary.bloodpressures.SeeAllClicked
 import org.simple.clinic.util.UserClock
 import org.simple.clinic.util.UtcClock
 import org.simple.clinic.util.toLocalDateAtZone
