+++ app/src/main/java/org/simple/clinic/summary/bloodpressures/view/NewBloodPressureSummaryViewInjector.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.summary.bloodpressures.newbpsummary.view
+package org.simple.clinic.summary.bloodpressures.view
 
 interface NewBloodPressureSummaryViewInjector {
   fun inject(target: NewBloodPressureSummaryView)
