+++ app/src/test/java/org/simple/clinic/summary/bloodpressures/NewBloodPressureSummaryViewEffectHandlerTest.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.summary.bloodpressures.newbpsummary
+package org.simple.clinic.summary.bloodpressures
 
 import com.nhaarman.mockito_kotlin.doReturn
 import com.nhaarman.mockito_kotlin.mock
@@ -12,6 +12,15 @@ import org.junit.Test
 import org.simple.clinic.bp.BloodPressureRepository
 import org.simple.clinic.mobius.EffectHandlerTestCase
 import org.simple.clinic.patient.PatientMocker
+import org.simple.clinic.summary.bloodpressures.BloodPressuresCountLoaded
+import org.simple.clinic.summary.bloodpressures.BloodPressuresLoaded
+import org.simple.clinic.summary.bloodpressures.LoadBloodPressures
+import org.simple.clinic.summary.bloodpressures.LoadBloodPressuresCount
+import org.simple.clinic.summary.bloodpressures.NewBloodPressureSummaryViewEffectHandler
+import org.simple.clinic.summary.bloodpressures.NewBloodPressureSummaryViewUiActions
+import org.simple.clinic.summary.bloodpressures.OpenBloodPressureEntrySheet
+import org.simple.clinic.summary.bloodpressures.OpenBloodPressureUpdateSheet
+import org.simple.clinic.summary.bloodpressures.ShowBloodPressureHistoryScreen
 import org.simple.clinic.util.scheduler.TrampolineSchedulersProvider
 import java.util.UUID
